refactor: split mixins into modules
