+++ pkgs/standards/autoapi/autoapi/v3/orm/mixins/markers.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+from ...types import declarative_mixin
+
+
+@declarative_mixin
+class AsyncCapable:
+    pass
+
+
+@declarative_mixin
+class Audited:
+    pass
+
+
+__all__ = ["AsyncCapable", "Audited"]
